requirements.txt
Faker==35.2.0
openpyxl==3.1.5
pytest==8.3.4
pytest-html-reporter==0.2.9
pytest-xdist==3.6.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.28.1
selenium-page-factory==2.7
webdriver-manager==4.0.2
six==1.17.0
pytest==8.3.4
pytest-html==4.1.1




